correct names of RCs, navbar css tweaked for small screens
