--- client/src/components/Navbar.js
@@ -1,58 +1,0 @@
-import { Link } from "react-router-dom";
-import { useLogOut, useAuthContext } from "../hooks";
-
-const Navbar = () => {
-  const { user } = useAuthContext();
-  const { logOut } = useLogOut();
-
-  const handleClick = () => logOut();
-
-  return (
-    <header>
-      <div className="container">
-        <Link to="/">
-          <h1>
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              width="24"
-              height="24"
-              viewBox="0 0 24 24"
-              fill="none"
-              stroke="currentColor"
-              strokeWidth="2"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              className="icon icon-tabler icons-tabler-outline icon-tabler-barbell"
-            >
-              <path stroke="none" d="M0 0h24v24H0z" fill="none" />
-              <path d="M2 12h1" />
-              <path d="M6 8h-2a1 1 0 0 0 -1 1v6a1 1 0 0 0 1 1h2" />
-              <path d="M6 7v10a1 1 0 0 0 1 1h1a1 1 0 0 0 1 -1v-10a1 1 0 0 0 -1 -1h-1a1 1 0 0 0 -1 1z" />
-              <path d="M9 12h6" />
-              <path d="M15 7v10a1 1 0 0 0 1 1h1a1 1 0 0 0 1 -1v-10a1 1 0 0 0 -1 -1h-1a1 1 0 0 0 -1 1z" />
-              <path d="M18 8h2a1 1 0 0 1 1 1v6a1 1 0 0 1 -1 1h-2" />
-              <path d="M22 12h-1" />
-            </svg>
-            FitTrack
-          </h1>
-        </Link>
-        <nav>
-          {user && (
-            <div>
-              <span>{user.email}</span>
-              <button onClick={handleClick}>Log Out</button>
-            </div>
-          )}
-          {!user && (
-            <div>
-              <Link to="/login">Log In</Link>
-              <Link to="/signup">Sign Up</Link>
-            </div>
-          )}
-        </nav>
-      </div>
-    </header>
-  );
-};
-
-export { Navbar };
