+++ client/src/components/WorkoutDetails.jsx
