+++ client/src/icons/index.js
@@ -1,0 +1,3 @@
+import { Dumbbell } from "./Dumbbell";
+
+export { Dumbbell };
