+++ client/src/index.css
@@ -28,6 +28,14 @@ header .container {
   justify-content: space-between;
 }
 
+.fitTrack {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 100%;
+  font-size: 2em
+}
+
 header a {
   color: #f1f1f1;
   text-decoration: none;
@@ -52,6 +60,8 @@ nav {
 
 nav a {
   margin-left: 10px;
+  font-weight: bold;
+  font-size: 1.2em
 }
 
 nav button {
@@ -120,6 +130,22 @@ form.login {
   .home {
     grid-template-columns: 1fr;
     gap: 40px;
+  }
+}
+
+@media (max-width: 420px) {
+.fitTrack,
+  nav a {
+    font-size: .9rem;
+  }
+}
+
+@media (max-width: 300px) {
+  .fitTrack, .container,
+  nav a {
+    font-size: .8rem;
+    margin-left: 4px;
+    padding: 1px 
   }
 }
